feat(apollo-vertex): simplify logic
