+++ apps/apollo-vertex/registry.json
@@ -266,11 +266,6 @@
       "registryDependencies": [],
       "files": [
         {
-          "path": "registry/ai-chat/types.ts",
-          "type": "registry:lib",
-          "target": "components/ui/ai-chat/types.ts"
-        },
-        {
           "path": "registry/ai-chat/adapters/agenthub/types.ts",
           "type": "registry:lib",
           "target": "components/ui/ai-chat/adapters/agenthub/types.ts"
@@ -301,16 +296,6 @@
           "target": "components/ui/ai-chat/hooks/use-sticky-scroll.ts"
         },
         {
-          "path": "registry/ai-chat/charts/chart-shared-types.ts",
-          "type": "registry:lib",
-          "target": "components/ui/ai-chat/charts/chart-shared-types.ts"
-        },
-        {
-          "path": "registry/ai-chat/charts/table-chart-types.ts",
-          "type": "registry:lib",
-          "target": "components/ui/ai-chat/charts/table-chart-types.ts"
-        },
-        {
           "path": "registry/ai-chat/charts/table-data-model.ts",
           "type": "registry:lib",
           "target": "components/ui/ai-chat/charts/table-data-model.ts"
@@ -339,16 +324,6 @@
           "path": "registry/ai-chat/tools/choices/index.ts",
           "type": "registry:lib",
           "target": "components/ui/ai-chat/tools/choices/index.ts"
-        },
-        {
-          "path": "registry/ai-chat/tools/table/table-tool.tsx",
-          "type": "registry:ui",
-          "target": "components/ui/ai-chat/tools/table/table-tool.tsx"
-        },
-        {
-          "path": "registry/ai-chat/tools/table/index.ts",
-          "type": "registry:lib",
-          "target": "components/ui/ai-chat/tools/table/index.ts"
         },
         {
           "path": "registry/ai-chat/tools/data-fabric-table/data-fabric-table-tool.tsx",
