--- apps/apollo-vertex/registry/ai-chat/charts/chart-shared-types.ts
@@ -1,6 +1,0 @@
-export type ColumnType = "numeric" | "string" | "boolean" | "datetime";
-
-export interface Column {
-  name: string;
-  type: ColumnType;
-}
