--- apps/apollo-vertex/registry/ai-chat/charts/table-chart-types.ts
@@ -1,6 +1,0 @@
-export interface TableChartConfig {
-  id: string;
-  name: string;
-  type: "table";
-  dimensions: string[];
-}
