+++ apps/apollo-vertex/registry/ai-chat/tools/data-fabric-table/data-fabric-table-tool.tsx
@@ -1,7 +1,33 @@
 "use client";
 
 import { toolDefinition } from "@tanstack/ai";
-import { tableInput, renderTable, type TableToolRenderOptions } from "../table";
+import { Suspense } from "react";
+import { z } from "zod";
+import { NoDataMessage } from "../../charts/no-data-message";
+import { buildStaticDataAdapter } from "../../charts/static-data-adapter";
+import { TableChartCard } from "../../charts/table-chart-card";
+import {
+  buildTableConfiguration,
+  buildTableDataModel,
+} from "../../charts/table-data-model";
+
+const columnSchema = z.object({
+  name: z.string().describe("Column name"),
+  type: z
+    .enum(["string", "numeric", "boolean", "datetime"])
+    .default("string")
+    .describe("Data type"),
+});
+
+const tableInput = z.object({
+  title: z.string().optional().describe("Table title"),
+  columns: z.array(columnSchema).min(1).describe("Column definitions"),
+  rows: z
+    .array(z.record(z.string(), z.unknown()))
+    .describe("Data rows as objects keyed by column name"),
+});
+
+type TableInput = z.infer<typeof tableInput>;
 
 export const DATA_FABRIC_TABLE_PROMPT = `You have a "data_fabric_table" tool.
 When the user wants to see data as a table, call it with column definitions and row data.
@@ -19,9 +45,22 @@ export const dataFabricTableClient = dataFabricTableDef.client(
   (input) => input,
 );
 
-export function renderDataFabricTable(
-  output: unknown,
-  options?: TableToolRenderOptions,
-) {
-  return renderTable(output, options);
+export function renderDataFabricTable(output: unknown) {
+  const { columns, rows, title } = output as TableInput;
+  if (!columns?.length || !rows) {
+    return <NoDataMessage />;
+  }
+  const dataModel = buildTableDataModel(columns);
+  const configuration = buildTableConfiguration(columns, title);
+  const cacheKey = `${rows.length}:${columns.map((c) => c.name).join(",")}`;
+  const adapter = buildStaticDataAdapter(rows, cacheKey);
+  return (
+    <Suspense>
+      <TableChartCard
+        configuration={configuration}
+        dataModel={dataModel}
+        dataAdapter={adapter}
+      />
+    </Suspense>
+  );
 }
