--- apps/apollo-vertex/registry/ai-chat/tools/table/index.ts
@@ -1,7 +1,0 @@
-export {
-  columnSchema,
-  tableInput,
-  type TableInput,
-  type TableToolRenderOptions,
-  renderTable,
-} from "./table-tool";
