--- apps/apollo-vertex/registry/ai-chat/types.ts
@@ -1,2 +1,0 @@
-export type { ChoiceOption } from "./tools/choices";
-export type { ToolsRenderer } from "./components/ai-chat-message";
